chore(cli): adds `--force-exit` command to CLI test
Aaron suggested this as a "hacky" way to prevent the CLI test run to fail on CI.
The error we're trying to prevent is:

```sh
Jest did not exit one second after the test run has completed.

'This usually means that there are asynchronous operations that weren't stopped in your tests. Consider running Jest with `--detectOpenHandles` to troubleshoot this issue.
```