feat: `--useMultiSig` support for governance proposals (#233)
* feat(cli/governance:propose): adds MultiSig.sol support

Also refactors checks to account for possible multisig proposer.

* chore(cli/governance:propose): updates CLI docs example

1. Increases `--deposit` from `10000` to `10000e18` to match current minimum (10,000 CELO)
2. Adds example with `--useMultiSig`

* style(cli/governance:approve): linting

* feat(cli/governance:propose): adds `--for` multisig flag to fix bug

* style(cli/governance:approve): linting

* chore(contractkit+cli): fix `isOwner` camel-case function name

* docs(contractkit): adds docs with `isOwner` fix

* test(cli/governance:propose): adds WIP test with TODO

* test(cli/governance:propose): adds MVP multisig creation (n=1)

Test passes, but needs to be significantly refactored.
1. there are various debugging statements
2. huge bytecode without source or comment
3. should probably live in a `beforeAll` so it can be re-used

* test(cli/governance:propose): refactors multisig creation

Now in a `beforeAll` and bytecode in `constants.ts`

* test(cli/governance:propose): refactors multisig creation into a function

* test(cli/governance:propose): adds test for multisig with 2 signers

* test(cli/governance:propose): remove `TODO` comments

* feat(cli/governance:propose): adds descriptive error message

The error messages are more descriptive, when the `--for` or `--usemultisig` flags are not set correctly.

* chore(cli): adds changeset for cli change

* chore(contractkit): adds changeset for drive-by contractkit change

* chore(cli/governance:propose): removes TODO comment

* chore(cli/governance:propose): improves `--for` flag error message

* chore(cli/test-utils): replaces `parseInt` with `BigInt`

`parseInt` converts to a number, which is not really safe when dealing with the large numbers we have on the EVM. Instead converting to a `BigInt` ensures better number handling.

* chore(cli): adds `--force-exit` command to CLI test

Aaron suggested this as a "hacky" way to prevent the CLI test run to fail on CI.
The error we're trying to prevent is:

```sh
Jest did not exit one second after the test run has completed.

'This usually means that there are asynchronous operations that weren't stopped in your tests. Consider running Jest with `--detectOpenHandles` to troubleshoot this issue.
```

* test(cli/governance:propose): removes `beforeAll` to fix snapshot error

`beforeAll` interfered with `snapId` as defined in `testWithGanache`.
As a short-term fix, we agreed to remove `beforeAll` and create multisigs on-demand in the specific `test()`.

When we migrate from Ganache to Anvil, we might consider adding better support for `beforeAll` without interfering with snapshots.