Update actions/cache to v4 (#346)
### Description

This PR aims to speed up the overall build time by switching cache
action version to v4. The problem with previous version was that post
restore steps were unexpectedly slow. On the new version this does not
seem to be a problem anymore.

### Other changes

None.

### Tested

Ran CI multiple times.

<!-- start pr-codex -->

---

## PR-Codex overview
This PR updates the version of the `actions/cache` used in the GitHub
Actions workflow from `v3` to `v4` for both the node cache and the build
artifacts cache.

### Detailed summary
- Updated `actions/cache` version from `v3` to `v4` for the `Restore
node cache` step.
- Updated `actions/cache` version from `v3` to `v4` for the `Restore
build artifacts cache` step.

> ✨ Ask PR-Codex anything about this PR by commenting with `/codex {your
question}`

<!-- end pr-codex -->