fix legacy ledger transfer / known token interactions.
i can only test on legacy ledger which seems to require there to NOT be a prefix for the token buffer. i believe but cant veerify that the modern version DOES want an 0x prefix

Added a test for wallet ledger.

the cip66 started failing but really that one was never implemented so we should skip it.

there was also an issue with the warning given on transfers. stableCurrency was called the gasCurrency which is wrong its the token being transferd