Fixing Ledger 1.1.10 interacting with whitelisted fee token contracts (#395)
There was a bug where ledgers would return a INVALID DATA error when
trying to transfer or interactive with a contract of one of the
whitelisted gas fee tokens. or using a gas token to pay for gas. This
occured while checking if the token was known in order to display human
readble data while confirming on the ledger.

The legacy 1.1.10 version of ledger app uses an older blob of signed
data. it seems in this case the encoded data should NOT be prefixed by
0x. while in the newer it should be.

this was not obvious and as automated tests are difficult when
interacting with hw devices was missed.

turned off test for cip66. Possibly we will remove later.

fixes #394
fixes #353

<!-- start pr-codex -->

---

This PR focuses on fixing issues related to token transfers and gas
currency handling in the `celo` ecosystem, particularly with the
`@celo/celocli` and `@celo/wallet-ledger` packages. It also improves
error messages and adds a new method for providing token information.

- Updated error messages in `packages/cli/src/transfer-stable-base.ts`
for clarity on transfer affordability.
- Fixed token information handling in
`packages/sdk/wallets/wallet-ledger/src/ledger-signer.ts`.
- Added `provideERC20TokenInformation` method to `LedgerSigner`.
- Improved tests in
`packages/sdk/wallets/wallet-ledger/src/ledger-wallet.test.ts` with mock
implementations and inline snapshots.
- Updated test suite to skip certain tests related to `cip66`.

> ✨ Ask PR-Codex anything about this PR by commenting with `/codex {your
question}`

<!-- end pr-codex -->