Backport ledger fix #395 (#396)
This is a backport of (#395)

There was a bug where ledgers would return a INVALID DATA error when
trying to transfer or interactive with a contract of one of the
whitelisted gas fee tokens. or using a gas token to pay for gas. This
occured while checking if the token was known in order to display human
readable data while confirming on the ledger.

The legacy 1.1.10 version of ledger app uses an older blob of signed
data. it seems in this case the encoded data should NOT be prefixed by
0x. while in the newer it should be.

this was not obvious and as automated tests are difficult when
interacting with hw devices was missed.

turned off test for cip66. Possibly we will remove later.

fixes #394
fixes #353





<!-- start pr-codex -->

---

## PR-Codex overview
This PR focuses on fixing issues related to token transactions and gas
currency handling in the `@celo/celocli` and `@celo/wallet-ledger`
packages, improving transaction handling and error messages.

### Detailed summary
- Fixed incorrect messaging for transferred tokens in the
`@celo/celocli`.
- Improved handling of token transfers and gas currency with ledger
devices in the `@celo/celocli`.
- Resolved issues with fee token transactions on ledger firmware 1.1.10
in the `@celo/wallet-ledger`.
- Updated error messages in `TransferStableBase` for clarity.
- Enhanced `provideERC20TokenInformation` method in `LedgerSigner` to
handle token info correctly.
- Modified tests to mock ledger interactions and ensure expected
behavior.

> ✨ Ask PR-Codex anything about this PR by commenting with `/codex {your
question}`

<!-- end pr-codex -->