Proposals and governance:build-proposal command tests and fixes (#421)
### Description

We had zero tests for building proposals. This adds them, switches the
default abis to be v12 and also fixes some contracts not loading when
building interactively since they are in sub directories.

#### Other changes

* move InteractiveProposalBuilder and ProposalBuilder to separate files.
This makes the PR look bigger regrettably. but only one small change was
made to these classes

* sets up jest testing for governance package 

### Tested

as well as the brand new tests I tests this by hand with several
contracts.

### Issues

- Fixes #413

<!-- start pr-codex -->

---

## PR-Codex overview
This PR focuses on updating the `@celo/governance` and `@celo/celocli`
packages, enhancing the proposal building features, fixing compatibility
issues with contracts, and improving testing configurations. It also
updates dependencies, particularly the `inquirer` library.

### Detailed summary
- Patches for `@celo/governance` and `@celo/celocli`.
- Fixes proposal building with contracts from `mento` and Solidity 0.8.
- Updates `inquirer` library version from `7.0.5` to `7.3.3`.
- Modifies `jest.config.js` for test matching patterns.
- Introduces global setup in `jest_setup.ts`.
- Adds improved transaction prompting in `InteractiveProposalBuilder`.
- Updates `ProposalBuilder` for better transaction handling.
- Adds tests for governance proposal building and interaction.

> ✨ Ask PR-Codex anything about this PR by commenting with `/codex {your
question}`

<!-- end pr-codex -->