DerivationPath Alias (#471)
### Description

add alias for celoLegacy to derivationPath flag on account:new command

This is currently same behavior as the default but that will change in a
future version.

#### Other changes

shift imports to be direct from celo/base instead of reexported from
cryptographic utils

### Related issues

- part of #448

<!-- start pr-codex -->

---

## PR-Codex overview
This PR introduces enhancements to the `account:new` command in the Celo
CLI, including new derivation path aliases and improved output for the
BIP44 derivation path. It also updates documentation and ensures better
error handling for invalid derivation paths.

### Detailed summary
- Added `DerivationPathAliases` enum with `eth` and `celoLegacy`.
- Updated `account:new` command to support new derivation path aliases.
- Improved output to include full BIP44 derivation path.
- Updated documentation for `--derivationPath` flag.
- Enhanced error handling for invalid derivation paths.

> ✨ Ask PR-Codex anything about this PR by commenting with `/codex {your
question}`

<!-- end pr-codex -->