Support delegation for vote signers (#508)
### Description

This PR adds support for delegating locked gold as a vote signer. 

#### Other changes

Added a test that covers a case when vote signer is authorized for a
`ReleaseGold` contract instance.

### Tested

Ran tests locally.

<!-- start pr-codex -->

---

## PR-Codex overview
This PR introduces support for delegation in the voting process for vote
signers within the `lockedgold` command of the CLI, enhancing governance
capabilities.

### Detailed summary
- Updated the `delegate.ts` file to include `address` in the
`newCheckBuilder`.
- Changed validation from `.isAccount(address)` to
`.isVoteSignerOrAccount()`.
- Added a new test case in `delegate.test.ts` for delegating as a vote
signer.
- Included deployment of a `releaseGoldContract` in the test setup.
- Verified delegation functionality and ensured the `delegateeAddress`
is recorded correctly.

> ✨ Ask PR-Codex anything about this PR by commenting with `/codex {your
question}`

<!-- end pr-codex -->