fix network:contracts.  (#514)
### Description

running `celocli network:contracts` would lead to a cryptic error 

```

 Error: execution reverted
    at decoratedCallback (/Users/martinvol/celo/developer-tooling/packages/sdk/connect/lib/utils/rpc-caller.js:109:23)
    at /Users/martinvol/celo/developer-tooling/node_modules/web3-core/node_modules/web3-providers-http/lib/index.js:131:13
    at process.processTicksAndRejections (node:internal/process/task_queues:95:5)
    Error: execution reverted
```

adding try/catch around some calls lead to the discovery that the call
to fetch the version number was failing for GovernanceSlasher contract.

A chat with @martinvol revealed there was an issue and we should display
version as 1.1.0.0 if none could be fetched.

#### Other changes


### Tested

added to and improve the contracts-l2 tests

### How to QA

run. `celocli network:contracts` command and expect to see a table to
contracts with versions

### Related issues

<!-- start pr-codex -->

---

## PR-Codex overview
This PR focuses on fixing a bug related to the `GovernanceSlasher`
contract's version retrieval, enhancing error handling when fetching
contract information, and updating tests to accommodate these changes.

### Detailed summary
- Fixed bug in version retrieval for `GovernanceSlasher`.
- Added error handling in `contracts.ts` when fetching contract version.
- Updated test cases in `contracts-l2.test.ts` to handle scenarios where
version retrieval fails.
- Mocked `ICeloVersionedContract` for better test isolation.
- Adjusted snapshot tests to reflect new behavior and outputs.

> ✨ Ask PR-Codex anything about this PR by commenting with `/codex {your
question}`

<!-- end pr-codex -->