use latest abis (#517)
### Description

ensure latest abis are used and that the same version is used on both
celocli and governance pacakges

<!-- start pr-codex -->

---

## PR-Codex overview
This PR focuses on updating the `@celo/abis-12` dependency across
various packages to the latest version, `12.0.0-canary.86`, and includes
minor adjustments in the `yarn.lock` file to reflect these changes.

### Detailed summary
- Updated `@celo/abis-12` from `12.0.0-canary.82` to `12.0.0-canary.86`
in:
  - `packages/sdk/contractkit/package.json`
  - `packages/sdk/governance/package.json`
- Updated `yarn.lock` to reflect the new version `12.0.0-canary.86`.
- Removed references to previous versions `12.0.0-canary.76` and
`12.0.0-canary.82`.

> ✨ Ask PR-Codex anything about this PR by commenting with `/codex {your
question}`

<!-- end pr-codex -->