feat: add additional checks to the governance:propose descriptionURL flag (#540)
This PR fixes:
- #509

by adding a check to the `descriptionURL` flag, it now **must** start
with `https://github.com/celo-org/governance/blob/main/CGPs` to be
proposed

<!-- start pr-codex -->

---

## PR-Codex overview
This PR introduces a validation check for the `descriptionURL` flag in
the `governance:propose` command to ensure it points to the correct
governance repository. It updates related documentation and tests
accordingly.

### Detailed summary
- Added `parseProposalDescriptionURL` function to validate
`descriptionURL`.
- Updated `CustomFlags` to include `proposalDescriptionURL`.
- Changed `descriptionURL` in `governance:propose` to use
`proposalDescriptionURL`.
- Updated command examples and documentation to reflect new URL format.
- Modified tests to check for valid and invalid `descriptionURL`.

> ✨ Ask PR-Codex anything about this PR by commenting with `/codex {your
question}`

<!-- end pr-codex -->