refactor: trigger upload with workflow_call (#560)
As figured out in
https://github.com/orgs/community/discussions/25281#discussioncomment-3300251

the current workflow uploading executables cannot work, it just won't be
called to prevent worfklow recursion.

Knowing that, I refactored the workflow to use `workflow_call` which is
called in the `release.yml`

<!-- start pr-codex -->

---

## PR-Codex overview
This PR introduces a Perl script to assert the stability of the
`celocli` version based on Semantic Versioning (SemVer) and updates the
GitHub Actions workflow to incorporate this check during the release
process.

### Detailed summary
- Added a new script `scripts/assert_stable_version` to validate
`celocli` versions against SemVer.
- Updated `.github/workflows/release.yaml` to include a step for
asserting stable release.
- Modified `upload-celocli-executables.yml` to assert stable version
before uploading.
- Enhanced version parsing in `upload-celocli-executables.yml` to use
`package.json`.
- Corrected file path in `git add` command for the formula file.

> ✨ Ask PR-Codex anything about this PR by commenting with `/codex {your
question}`

<!-- end pr-codex -->