ci: big refactor to handle platform specific node native modules (#582)
Initially the bug this PR was trying to fix was just allow he usage of
the `--useLedger` flag for homebrew users, but it turns out it's not
only `node-hid` which was affected but any native c/c++ module that can
be deep down our node_modules.

To prevent future issues, it uses different runner to compile and pack
the archives on their own specific platform (windows, linux, linux-arm,
etc..).

I triggered a flow here
https://github.com/celo-org/developer-tooling/actions/runs/14265200448
and it successfully built the executables for each platform

<!-- start pr-codex -->

## PR-Codex overview
This PR introduces a new job in the GitHub Actions workflow for
uploading `celocli` executables. It enhances variable setup, improves
dependency management, and modifies how the release tarball is handled
and uploaded.

### Detailed summary
- Added `setupVariables` job for version and commit extraction.
- Simplified `compile` job with matrix strategy for OS targets.
- Changed tarball handling to download from GitHub releases.
- Updated variable references to use outputs from `setupVariables`.
- Improved Homebrew PR preparation and commit message formatting.

> ✨ Ask PR-Codex anything about this PR by commenting with `/codex {your
question}`

<!-- end pr-codex -->