new way to keep generated docs up to date (#585)
### Description

This solves 2 issues. 

1. when links to docs are made it always points to the latest version of
the repo rather than the version that was released, so now use the git
tag for the link instead of branch name
2. for us devs its a little bit annoying how easy it is to forget to run
yarn docs for a pr. and really there is no reason they need to be
updated each pr anyway. so instead add a command that runs after the
packages are versioned which will generated new docs for that version,

#### Other changes

i plan that once this is in we turn off the doc build requirement on prs

### Tested

n/a

### How to QA

unsure.

<!-- start pr-codex -->

---

## PR-Codex overview
This PR updates the documentation links and modifies the `package.json`
scripts for the `celocli` project to reflect a new versioning scheme and
improve the release process.

### Detailed summary
- Updated documentation links to point to the new version
`@celo/celocli@6.2.5`.
- Changed `package.json` script `version-and-reinstall` to
`version-then-update-files` to include `yarn build` and `yarn docs`.
- Modified `release` workflow name to "Open Release PR or Publish
Packages" in `.github/workflows/release.yaml`.
- Adjusted `version` in the release workflow to use the new script.

> The following files were skipped due to too many changes:
`docs/command-line-interface/account.md`

> ✨ Ask PR-Codex anything about this PR by commenting with `/codex {your
question}`

<!-- end pr-codex -->