Fix admin revoke command for L2 (#601)
### Description

admin revoke was using the personal name space as a way to set the
authorized voted signer to a meaningless address, since personal is now
deprecated instead set it back to the default account.

#### Other changes

since this was the last cli test using ganache remove the ganache cli
test from ci

### Tested


### How to QA

_List of steps required to QA the change if applicable_

### Related issues

- Fixes #257 #539

<!-- start pr-codex -->

---

## PR-Codex overview
This PR focuses on fixing the `releasecelo/admin-revoke` command to
ensure it works correctly on L2, along with updates to test
configurations and dependencies.

### Detailed summary
- Updated `releasecelo/admin-revoke` to work with L2.
- Changed `contractAddress` return type to `StrongAddress`.
- Modified test configurations in `package.json` and
`.github/workflows/ci.yml`.
- Updated `gitguardian.yml` to reflect a new test key.
- Enhanced tests in `admin-revoke.test.ts` for better coverage and
functionality.

> ✨ Ask PR-Codex anything about this PR by commenting with `/codex {your
question}`

<!-- end pr-codex -->