election:current on viem (#604)
### Description

converts this highly used command to viem


#### Other changes

* setting batch on the viem client reduces number of queries to rpc as
it automatically creates multicall calls when promise.all is used.


### Tested

current tests so far, once we agree on how packages to be should be
tests should be created on that external api too

### How to QA

run it locally

### Related issues

- Fixes #603

<!-- start pr-codex -->

---

## PR-Codex overview
This PR focuses on enhancing the `multicall` functionality within the
CLI, updating the governance commands and election processes to utilize
`multicall` where applicable. It also modifies the output for elected
nodes in the election command.

### Detailed summary
- Added `batch: { multicall: true }` to `createPublicClient`.
- Implemented `deployMultiCall` function for setting up multicall in
tests.
- Updated election command to fetch and display elected community RPC
nodes.
- Refactored output tables for elected nodes to include new properties.
- Modified governance commands to utilize `multicall` in tests.
- Changed descriptions and comments for clarity regarding multicall
usage.

> ✨ Ask PR-Codex anything about this PR by commenting with `/codex {your
question}`

<!-- end pr-codex -->