Migrate RPC and Validator Lists to viem (#611)
### Description

* validator:list and validator:rpc-urls both now use viem
* refactored rpc-urls command to take better advantange of automatic
multicall with viem

#### Other changes


* I added to get public client that we are connecting to he nodeURl. I
think its nice feedback that something is happening and that the url is
as expected
* Started to call validators as Community RPC Nodes in places where it
doesnt really change the output interface


### Tested

ran commands locally on a few chains and also the tests pass

### How to QA

run the commands 

### Related issues

- Fixes #608

<!-- start pr-codex -->

---

## PR-Codex overview
This PR focuses on migrating various validator-related commands and
functionalities to use the `viem` library, optimizing performance, and
enhancing the user experience by improving how validators and their
metadata are fetched and displayed.

### Detailed summary
- Migrated `validator:list` and `validator:rpc-urls` commands to use
`viem`.
- Enhanced fetching of registered validators and their group
affiliations.
- Improved concurrency level for fetching data.
- Added new functions for retrieving validator group and metadata URLs.
- Updated user feedback messages using `ux.action`.

> ✨ Ask PR-Codex anything about this PR by commenting with `/codex {your
question}`

<!-- end pr-codex -->