Deprecate DKG commands and useAKV flag (#615)
### Description

We are hiding the dkg flags from documentation and marking them as
deprecated.


mark useAKV option as deprecated. 


note that these deprecation is reversable.

### Tested

no

### How to QA


### Related issues

- Fixes #370

<!-- start pr-codex -->

---

## PR-Codex overview
This PR focuses on deprecating the `DKG` command group and related flags
in the `celocli` package, signaling a transition away from the
Decentralized Key Generation functionality.

### Detailed summary
- Deleted `dkg.md.changeset/chatty-bananas-tell.md`.
- Marked the `DKG` command group as deprecated.
- Added deprecation options in `packages/cli/src/utils/notice.ts`.
- Updated descriptions in `package.json` for clarity.
- Added `hidden` and `deprecationOptions` to several `DKG` command
classes.
- Deprecated the `useAKV` flag in the context of Azure KeyVault.

> ✨ Ask PR-Codex anything about this PR by commenting with `/codex {your
question}`

<!-- end pr-codex -->