chore: account:new cleanup (#628)
This PR initially aimed to migrate as much as possible to viem, but it
turns out this is a bbigger undertaking as it requires refactor
@celo/base and @celo/cryptographic-utils . The changes were rolled back
and this is a very minimal refresh

- Fixes #623

<!-- start pr-codex -->

---

## PR-Codex overview
This PR focuses on minor improvements and updates to type definitions,
package versions, and output formatting in the `@celo` ecosystem,
enhancing the overall code quality and user experience.

### Detailed summary
- Updated `viem` package version to `^2.23.5` in `package.json` files.
- Improved type definition for `isHexString` function in `address.ts`.
- Removed `accountAddress` from output in `new.ts` and its tests to
streamline the output.

> ✨ Ask PR-Codex anything about this PR by commenting with `/codex {your
question}`

<!-- end pr-codex -->