refactor account:list --use ledger to no longer need --node flag (#640)
### Description

needing to pass a node flag to get the list of accounts on the ledger
was frustrating. The cause was that clients need a transport. but do we
need a client? actually no. just the list of ledger accounts.

#### Other changes
* refactored the `ledgerToWalletClient` so that the code for getting the
accounts could be called separately without creating a wallet client.
* removed node:accounts since it has been depreated for 5 years (based
on commit history of when the deprecation notice was written)

### Tested

new tests, old test, ran acount:list --useLedger 


### How to QA

run some commands with your ledger hooked up

### Related issues

- Fixes #161