Converts governance:vote to viem (#639)
### Description

viem is the future

#### Other changes

* typed `NULL_ADDRESS` as having 0x prefix. 

* added `type HexString` to celo/base

* fewer places use our Address type. 
* can pass a bigInt to a few places that are checking proposalID is
valid. just a type change needed since BigInt() takes bigint as an
argument already

* testIWithViem now accepts a forkUrl and forkBlockNumber. if passed the
anvil chain will be based on that state instead of loading state form
the devchain.

### Tested

just existing tests.

### How to QA

need to have a proposal to vote on. 

### Related issues

- Fixes #622

<!-- start pr-codex -->

---

## PR-Codex overview
This PR focuses on enhancing type safety and clarity in the governance
and voting functionalities within the Celo ecosystem. It introduces
explicit types for addresses, refines the vote proposal process, and
improves testing for voting functionalities.

### Detailed summary
- Added `HexString` type for hexadecimal strings.
- Exported `voteProposal`, `VoteProposalAdapter`, and
`VoteProposalTypes` from `./vote`.
- Refined address types to `StrongAddress` in various files.
- Updated `vote` function to use explicit types and handle voting logic.
- Added tests for `voteProposal` to ensure correct functionality.
- Enhanced governance command to utilize new voting logic and types.

> ✨ Ask PR-Codex anything about this PR by commenting with `/codex {your
question}`

<!-- end pr-codex -->