Docs setup for new packages (plus eastereggs) (#647)
### Description

generate nice docs for our new packages. The configuration here is a
little different from out other packages it will only generate docs for
functions that are exported and have a typedoc comment written for them.

#### Other changes

* remove semver from root (not needed(
* set jest and vitest as optional peer dependencies of dev-utils
(removes warning)

### Tested

in this economy?

### How to QA

dont worry about it
### Related issues



<!-- start pr-codex -->

---

## PR-Codex overview
This PR introduces updates to the `@celo/core` and `@Celo/Actions`
packages, including new function documentation, enhancements to the
voting mechanism, and adjustments to the configuration files for
improved dependency management and TypeDoc settings.

### Detailed summary
- Added version information for `@celo/core` and `@Celo/Actions`.
- Documented new functions: `vote`, `voteProposal`, and
`getElectedRpcNodes`.
- Updated `package.json` to include optional peer dependencies for
`jest` and `vitest`.
- Enhanced TypeDoc configurations in `typedoc.json` files.
- Improved documentation structure and examples in Markdown files.

> ✨ Ask PR-Codex anything about this PR by commenting with `/codex {your
question}`

<!-- end pr-codex -->