fix ci install issue with node-hid /node-gyp (#658)
### Description

make sure everything is on same node version and install works. 

basically on node 22 it just kept having issues with node-hid. i even
tried out setting node hid to a newer version via resolution and still
it couldnt compile (node-gyp of course) so downgraded the ci back to
node 20 to fix that. (but also fixed a few workflows on node 18 still to
be node 20

#### Other changes

we dont need to run the ci twice on changeset prs

### Tested

thats what ci is for

### How to QA
n/a
### Related issues

<!-- start pr-codex -->

---

## PR-Codex overview
This PR focuses on upgrading the Node.js version used in various GitHub
Actions workflows from version `18` and `22` to version `20`. It also
updates dependencies in the `yarn.lock` file.

### Detailed summary
- Updated Node.js version to `20` in:
  - `.github/workflows/changeset-for-renovate.yml`
  - `.github/actions/sync-workspace/action.yml`
  - `.github/workflows/release.yaml`
  - `.github/workflows/ci.yml`
- Incremented `NODE_MODULE_CACHE_VERSION` from `5` to `6`.
- Updated dependencies in `yarn.lock`:
  - Changed `@ledgerhq/devices` from version `8.2.2` to `8.4.5`.
  - Updated `@ledgerhq/errors` from version `6.16.3` to `6.21.0`.
- Updated `@ledgerhq/hw-transport-node-hid-noevents` from version
`6.30.5` to `6.30.6`.
  - Updated `@ledgerhq/hw-transport` from version `6.30.5` to `6.31.5`.

> ✨ Ask PR-Codex anything about this PR by commenting with `/codex {your
question}`

<!-- end pr-codex -->