enabled prod releasing (#701)
its time for full release since cr 13 is almost out out

<!-- start pr-codex -->

---

## PR-Codex overview
This PR focuses on updating the `mode` in the `.changeset/pre.json` file
and modifying tests related to the `releaseGoldWrapper` in
`withdraw.test.ts`, including changes in the time travel duration and
the expected values for the `currentReleasedTotal`.

### Detailed summary
- Changed `mode` from `"pre"` to `"exit"` in `.changeset/pre.json`.
- Updated the time travel duration from `DAY * 30` to `DAY * 31` in
`withdraw.test.ts`.
- Removed assertions for `currentReleasedTotal` before and after time
travel.
- Kept the assertion for `totalWithdrawn` unchanged.

> ✨ Ask PR-Codex anything about this PR by commenting with `/codex {your
question}`

<!-- end pr-codex -->