Require proposals description urls to return real url (#717)
### Description

it is a very big problem if the wrong url is submitted on chain. it will
cause mondo to break.

#### Other changes

n.a
### Tested

new tests

### How to QA

try to submited a proposals with a url that points to a cgp that does
not exist.

### Related issues

- Fixes #

<!-- start pr-codex -->

---

## PR-Codex overview
This PR introduces a validation check for the `descriptionURL` in the
`governance:propose` command, ensuring that the provided URL exists and
is reachable before allowing a proposal submission.

### Detailed summary
- Added a check to prevent submission of a non-existent
`descriptionURL`.
- Implemented `urlDestinationExists` method in `checks.ts` to verify URL
accessibility.
- Updated `Propose` command to include the new URL check in its
validation process.
- Added tests for scenarios where `descriptionURL` returns a non-200
status and when fetching the URL throws an error.

> ✨ Ask PR-Codex anything about this PR by commenting with `/codex {your
question}`

<!-- end pr-codex -->