breaking: Removes offchain-read and offchain-write commands. (#719)
### Description

Removes offchain-read and offchain-write commands.
These have been in deprecation for a year with no contestation. The
follow the deprecated cip8 spec.

#### Other changes

was able to also remove celo identity dependency
### Tested

removed code

### How to QA

n.a
### Related issues

- Fixes #222

<!-- start pr-codex -->

---

## PR-Codex overview
This PR focuses on the removal of deprecated commands related to
off-chain data in the `celo` CLI, specifically the `offchain-read` and
`offchain-write` commands. It also updates the `package.json` to reflect
the changes in dependencies.

### Detailed summary
- Deleted files: 
  - `packages/cli/src/utils/off-chain-data.ts`
  - `packages/cli/src/commands/account/offchain-read.ts`
  - `packages/cli/src/commands/account/offchain-write.ts`
- Updated `package.json`:
  - Removed `"@celo/identity": "^5.1.2"` from dependencies.
- Added a changeset detailing the removal of deprecated commands.

> ✨ Ask PR-Codex anything about this PR by commenting with `/codex {your
question}`

<!-- end pr-codex -->